@@ -18,7 +18,7 @@ import android.media.AudioManager
 import android.net.Uri
 import android.os.Build
 import android.preference.PreferenceManager.getDefaultSharedPreferences
-import android.support.v4.content.ContextCompat
+import androidx.core.content.ContextCompat
 import android.view.*
 import android.widget.SeekBar
 import android.widget.Toast
